+++ src/imageMeasure.js
@@ -26,7 +26,16 @@ ImageMeasure.prototype.measureImage = function (src) {
 		image = this.pdfKitDoc._imageRegistry[src];
 	}
 
-	return { width: image.width, height: image.height };
+	var imageSize = { width: image.width, height: image.height };
+
+	if (typeof image === 'object' && image.constructor.name === 'JPEG') {
+		// If EXIF orientation calls for it, swap width and height
+		if (image.orientation > 4) {
+			imageSize = { width: image.height, height: image.width };
+		}
+	}
+
+	return imageSize;
 
 	function realImageSrc(src) {
 		var img = that.imageDictionary[src];